moved processor to core module
