+++ dook/core/processor/errors.py
@@ -1,0 +1,4 @@
+NOTIFICATION_SENDING_FAILED_ERROR = "Sending notifications failed with exception: {}"
+NEWS_DRAFT_PROCESSING_WITH_UNHANDLED_EXCEPTION_ERROR = (
+    "Processing of {} failed with unhandled exception"
+)
