+++ dook/core/processor/managers.py
