--- dook/processor/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ProcessorConfig(AppConfig):
-    name = "processor"
